Switch from OSMDroid 4.2 to mapbox-android-sdk 0.3.0
This requires dropping support for Android API level 8.

Also removed the querying of MLS for position info.  That can be
added back in a separate activity, as this Map Activity should be
used by stumblers to determine where in their area has already been
stumbled.  The testing of MLS responses was never clear in the
MapActivity.

The "Refresh" button now refreshes the map back to the user's
location.  This is useful if they start to pan around the map a
bunch and get lost.